+++ src/main/kotlin/com/github/encryptsl/credit/api/events/CreditDepositEvent.kt
@@ -5,9 +5,10 @@ import org.bukkit.command.CommandSender
 import org.bukkit.event.Cancellable
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
+import java.math.BigDecimal
 
 @Suppress("UNUSED")
-class CreditDepositEvent(val commandSender: CommandSender, val offlinePlayer: OfflinePlayer, val money: Double, val silent: Boolean): Event(), Cancellable {
+class CreditDepositEvent(val commandSender: CommandSender, val offlinePlayer: OfflinePlayer, val money: BigDecimal, val silent: Boolean): Event(), Cancellable {
 
     private var isCancelled: Boolean = false
