+++ src/main/kotlin/com/github/encryptsl/credit/api/events/CreditWithdrawEvent.kt
@@ -4,9 +4,10 @@ import org.bukkit.command.CommandSender
 import org.bukkit.event.Cancellable
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
+import java.math.BigDecimal
 
 @Suppress("UNUSED")
-class CreditWithdrawEvent(val commandSender: CommandSender, val offlinePlayer: OfflinePlayer, val money: Double): Event(), Cancellable {
+class CreditWithdrawEvent(val commandSender: CommandSender, val offlinePlayer: OfflinePlayer, val money: BigDecimal): Event(), Cancellable {
 
     private var isCancelled: Boolean = false
