+++ src/main/kotlin/com/github/encryptsl/credit/api/events/GlobalCreditSetEvent.kt
@@ -4,9 +4,10 @@ import org.bukkit.command.CommandSender
 import org.bukkit.event.Cancellable
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
+import java.math.BigDecimal
 
 @Suppress("UNUSED")
-class GlobalCreditSetEvent(val commandSender: CommandSender, val money: Double): Event(), Cancellable {
+class GlobalCreditSetEvent(val commandSender: CommandSender, val money: BigDecimal): Event(), Cancellable {
 
     private var isCancelled: Boolean = false
